refactor: extract viewer serve logic from cli
Move viewer background/pidfile management into opencode_mem/commands/viewer_cmds.py and keep cli serve command as a thin wrapper.
